refactor(plugins): move PluginsErrorsInfo into a /plugins folder
